backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.3
flask-cors==3.0.10
python-dotenv==1.0.0